% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalibratePerformanceMeasure.R
\name{calibrateControlLimit}
\alias{calibrateControlLimit}
\title{Calibration of the control limit for the selected chart}
\usage{
calibrateControlLimit(targetARL = NULL, targetMRL = NULL, n, m,
  theta = NULL, Ftheta = NULL, dist, mu, sigma, dist.par = c(0, 1,
  1), chart, initial.par, replicates = 50000, isParallel = FALSE,
  maxIter = 20, progress = TRUE)
}
\arguments{
\item{targetARL}{is the target ARL to calibrate. By default is set to NULL}

\item{targetMRL}{is the target ARL to calibrate. By default is set to NULL}

\item{n}{is the subroup size}

\item{m}{is the reference sample size}

\item{theta}{is a constant. By default is set to NULL}

\item{Ftheta}{is a constant between (0,1). By default is set to NULL}

\item{dist}{is the distribution to analyze. By default is "Normal"}

\item{mu}{is a vector of two elements. By default is set to c(0,0)
The firt element refers to the mean of the reference sample.
The second element refers to the mean of the monitoring sample.}

\item{sigma}{is a vector of two elements. By default is set to c(1,1)
The firt element refers to the standard deviation of the reference sample.
The second element refers to the standard deviation of the monitoring sample.}

\item{dist.par}{is a vector of three elements. By defautl is set to c(0, 1, 1)
The firt element refers to location of the distribution
The second element refers to scale of the distribution
The third element refers to shape of the distribution}

\item{chart}{is the selected type of chart. Three options are available: Shewhart, CUSUM, EWMA}

\item{initial.par}{is a vector of one, or two elements depending
on the chart selected: Shewhart, CUSUM, EWMA
Shewhart chart is c(k), where k comes from UCL = mu + k*sigma, LCL = mu - k*sigma.
CUSUM chart is c(k, h, t) where k is the reference value and h is the control limit,
and t is the type of the chart (1:positive, 2:negative, 3:two sides)
EWMA chart is c(lambda, L), where lambda is the smoothing constant
and L multiplies standard deviation to get the control limit}

\item{replicates}{is a numeric and represent the number of replicates
to run the function getARL in each iteration}

\item{isParallel}{is a boolean. If is TRUE the code runs in parallel according to the
number of cores in the computer,otherwise the code runs sequentially. By default is set to TRUE}

\item{maxIter}{is a numeric. The maximum number of iteration to take the calibration before stops}

\item{progress}{is a boolean. If TRUE it shows the progress of the calibration in the console.}
}
\description{
The methodology used to calibrate the control limit
for the SNS chart depending on the chart selected
}
\examples{
n <- 2 # subgroup size
m <- 30 # reference-sample size
dist <- "Normal" # distribution
mu <- c(0, 0) # c(reference sample mean, monitoring sample mean)
sigma <- c(1, 1) # c(reference sample sd, monitoring sample sd)

#### Distribution parameters
dist.par <- c(0, 1, 1) # c(location, scale, shape)

#### Other Parameters
replicates <- 2
targetARL <- 370

#### Control chart parameters
chart <- "Shewhart"
chart.par <- c(3)
shewhart <- calibrateControlLimit(
  targetARL = targetARL, targetMRL = NULL, n = n, m = m, theta = NULL,
  Ftheta = NULL, dist = dist, mu = mu, sigma = sigma, dist.par = dist.par, initial.par = chart.par,
  replicates = replicates, chart = chart
)

chart <- "CUSUM"
chart.par <- c(0.5, 2.5, 3)
cusum <- calibrateControlLimit(
  targetARL = targetARL, targetMRL = NULL, n = n, m = m, theta = NULL,
  Ftheta = NULL, dist = dist, mu = mu, sigma = sigma, dist.par = dist.par, initial.par = chart.par,
  replicates = replicates, chart = chart
)

chart <- "EWMA"
chart.par <- c(0.2, 2.962)
ewma <- calibrateControlLimit(
  targetARL = targetARL, targetMRL = NULL, n = n, m = m, theta = NULL,
  Ftheta = NULL, dist = dist, mu = mu, sigma = sigma, dist.par = dist.par, initial.par = chart.par,
  replicates = replicates, chart = chart
)
}
