% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataAlignment.R
\name{dataAlignment}
\alias{dataAlignment}
\title{Alignment of the data}
\usage{
dataAlignment(X, Y, alignment = "unadjusted", constant = NULL)
}
\arguments{
\item{X}{is a numerical vector.}

\item{Y}{is a numerical vector.}

\item{alignment}{is the aligment of the data
unadjusted:
overallmean:
overallmedian:
samplemean:
samplemedian:
referencemean:
referencemedian:
constantvalue:}

\item{constant}{is a numeric value}
}
\description{
Align the monitoring sample X and the reference sample Y
}
\examples{
X = c(30, 45, 50)
Y = c(20, 22, 25, 30, 70)
dataAlignment(X,Y)

}
