% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examplesqiuDocument.R
\docType{data}
\name{example49}
\alias{example49}
\title{Data from Example 4.9 Qiu (2014).}
\format{A data frame with 100 rows and 2 columns:
\describe{
 \item{V1}{First data set with 100 observations divided in 20 batches of 5 observation each. The first 10 batches are N(0,1) and the second 10 batches are N(1,1).}
 \item{V2}{Second data set with 100 observations divided in 20 batches of 5 observation each. The first 10 batches are N(0,1) and the second 10 batches are N(0,22).}
}}
\source{
\url{http://users.phhp.ufl.edu/pqiu/research/book/spc/data/example49.dat}
}
\usage{
example49
}
\description{
A dataset containing the data set used in Example 4.9 of Qiu (2014).
}
\keyword{datasets}
