% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examplesqiuDocument.R
\docType{data}
\name{example81}
\alias{example81}
\title{Data from Example 8.1 on page 319 Qiu (2014).}
\format{A data frame with 30 rows and 11 columns. Each row is a batch. The first 10 columns are the observation of each batch:
\describe{
 \item{X}{1st observation of the batch.}
 \item{X.1}{2nd observation of the batch.}
 \item{X.2}{3rd observation of the batch.}
 \item{X.3}{4th observation of the batch.}
 \item{X.4}{5th observation of the batch.}
 \item{X.5}{6th observation of the batch.}
 \item{X.6}{7th observation of the batch.}
 \item{X.7}{8th observation of the batch.}
 \item{X.8}{9th observation of the batch.}
 \item{X.9}{10th observation of the batch.}
 \item{psi}{psi}
}}
\source{
\url{http://users.phhp.ufl.edu/pqiu/research/book/spc/data/example81.dat}
}
\usage{
example81
}
\description{
A dataset containing the data set used in Example 8.1 on page 319 of Qiu (2014).
}
\keyword{datasets}
