% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examplesqiuDocument.R
\docType{data}
\name{example82}
\alias{example82}
\title{Data from Example 8.2 Qiu (2014).}
\format{A data frame with 30 rows (batches) and 6 columns
\describe{
 \item{X}{first observation of the batch}
 \item{X.1}{second observation of the batch}
 \item{X.2}{third observation of the batch}
 \item{X.3}{fourth observation of the batch}
 \item{X.4}{fifth observation of the batch}
 \item{Xmedian}{Median of the batch}
}

(first 5 are the observations, the last is the median of the 5 observations).
The column names are the default names given when the dataframe was created.}
\source{
\url{http://users.phhp.ufl.edu/pqiu/research/book/spc/data/example82.dat}
}
\usage{
example82
}
\description{
A dataset containing the data set used in Example 8.2 on page 323 of Qiu (2014).
}
\keyword{datasets}
