% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examplesqiuDocument.R
\docType{data}
\name{example84}
\alias{example84}
\title{Data from Example 8.4 Qiu (2014).}
\format{A data frame with 30 rows (batches) and 7 columns:
\describe{
 \item{X}{first observation of the batch}
 \item{X.1}{second observation of the batch}
 \item{X.2}{third observation of the batch}
 \item{X.3}{fourth observation of the batch}
 \item{X.4}{fifth observation of the batch}
 \item{Wn}{Man Withney Statistic}
 \item{Cnn}{Cusum negative statistic}
 \item{Cnp}{Cusum positive statistic}
}}
\source{
\url{http://users.phhp.ufl.edu/pqiu/research/book/spc/data/example84.dat}
}
\usage{
example84
}
\description{
A dataset containing the data set used in Example 8.4 of Qiu (2014).
}
\keyword{datasets}
