% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examplesqiuDocument.R
\docType{data}
\name{example87}
\alias{example87}
\title{Data from Example 8.7 on page 339 Qiu (2014).}
\format{A data frame with 100 rows and 2 columns:
\describe{
 \item{x}{first 14 observations are in control}
 \item{Tmax}{test statistic for the method proposed by Hawkins and Deng (2010)}
}}
\source{
\url{http://users.phhp.ufl.edu/pqiu/research/book/spc/data/example87.dat}
}
\usage{
example87
}
\description{
A dataset containing the data set used in Example 8.7 on page 339 of Qiu (2014).
}
\keyword{datasets}
