% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ObtainAverageRunLength.R
\name{getARL}
\alias{getARL}
\title{Average Run Length (ARL)}
\usage{
getARL(n, m, theta = NULL, Ftheta = NULL, dist, mu, sigma,
  dist.par = c(0, 1, 1), chart.par, print.RL = FALSE,
  replicates = 10000, chart, progress = FALSE, isParallel = TRUE,
  calibrate = FALSE, arl0 = 370)
}
\arguments{
\item{n}{is the subroup size}

\item{m}{is the reference sample size}

\item{theta}{is a constant.}

\item{Ftheta}{is a constant between (0,1).}

\item{dist}{is a vector of characters. Choose one type:
"Normal":     Normal distribution. (Default)
"DoubleExp":  Double exponential distribution (also known as Laplace distribution).
"DoubleExp2": Double exponential squared distribution from a DoubleExp(0,1).
"LogNormal":  Lognormal distribution.
"Gamma":      Gamma distribution.
"Weibull":    Weibull distribution}

\item{mu}{user defined expected value.}

\item{chart.par}{is a vector of one, or two elements depending
on the chart selected:
\describe{
  \item{For Shewhart scheme}{is c(k), where k comes from UCL = mu + k*sigma, LCL = mu - k*sigma.}
  \item{For CUSUM scheme}{is c(k, h, t) where k is the reference value and h is the control limit,
  and t is the type of the chart (1:positive, 2:negative, 3:two sides)}
  \item{For EWMA scheme}{is c(lambda, L), where lambda is the smoothing constant
  and L multiplies standard deviation to get the control limit}
}}

\item{print.RL}{is boolean. If set TRUE the function getARLSNS return the vectors of RL for each iteration}

\item{replicates}{is a numeric and represent the number of replicates to get the ARL}

\item{chart}{is the selected type of chart. Three options are available: Shewhart, CUSUM, EWMA}

\item{progress}{is a boolean. If TRUE it shows the progress of the calibration in the console.}

\item{isParallel}{is a boolean. If is TRUE the code runs in parallel according to the
number of cores in the computer,otherwise the code runs sequentially. By default is set to TRUE}

\item{calibrate}{is a boolean. If is TRUE the RL is limit to 10 times the target ARL}

\item{arl0}{is a numeric value. Is the expected value of the RL, by default is set to 370}
}
\description{
Get the ARL \code{\link{getRL}}
}
\examples{
n <- 5 # subgroup size
m <- 100 # reference-sample size
dist <- "Normal"
mu <- c(0, 0) # c(reference sample mean, monitoring sample mean)
sigma <- c(1, 1) # c(reference sample sd, monitoring sample sd)

#### Distribution parameters
dist.par <- c(0, 1, 1) # c(location, scale, shape)

#### Other Parameters
replicates <- 2
print.RL <- TRUE
isParallel <- FALSE
calibrate <- FALSE
progress <- TRUE
arl0 <- 370

#### Control chart parameters
chart <- "Shewhart"
chart.par <- c(3)
shewhart <- getARL(n, m,
  theta = NULL, Ftheta = NULL, dist, mu, sigma, dist.par = dist.par,
  chart = chart, chart.par = chart.par, print.RL = print.RL, replicates = replicates, isParallel = isParallel,
  calibrate = calibrate, arl0 = arl0
)

chart <- "CUSUM"
chart.par <- c(0.25, 4.4181, 3)
cusum <- getARL(n, m,
  theta = NULL, Ftheta = NULL, dist, mu, sigma, dist.par = dist.par,
  chart = chart, chart.par = chart.par, print.RL = print.RL, replicates = replicates, isParallel = isParallel,
  calibrate = calibrate, arl0 = arl0
)

chart <- "EWMA"
chart.par <- c(0.2, 2.962)
shewhart <- getARL(n, m,
  theta = NULL, Ftheta = NULL, dist, mu, sigma, dist.par = dist.par,
  chart = chart, chart.par = chart.par, print.RL = print.RL, replicates = replicates, isParallel = isParallel,
  calibrate = calibrate, arl0 = arl0
)
}
