% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetDist.R
\name{getDist}
\alias{getDist}
\title{Random variate generetor}
\usage{
getDist(n, dist, mu, stdev, par.location = 0, par.scale = 1,
  par.shape = 1)
}
\arguments{
\item{n}{is a scalar. Indicates how many random variables to be generated}

\item{dist}{is a vector of characters. Choose one type:
"Normal":     Normal distribution. (Default)
"DoubleExp":  Double exponential distribution (also known as Laplace distribution).
"DoubleExp2": Double exponential squared distribution from a DoubleExp(0,1).
"LogNormal":  Lognormal distribution.
"Gamma":      Gamma distribution.
"Weibull":    Weibull distribution}

\item{mu}{user defined expected value.}

\item{stdev}{user defined standard deviation.}

\item{par.location}{scalar, location parameter, by defaoult is 0**.}

\item{par.scale}{scalar, scale parameter, by defaoult is 1**.}

\item{par.shape}{scalar, shape parameter, by defaoult is 1.
**Note on par.location and par.scale:
par.location and par.scale are the location and scale parameters
of the distribution that generates the desired distribution. For lognormal,
par.location and par.scale correspond to the location and scale parameters of the normal
distribution that generales the lognormal. Hence, in this case they are the logmean and
the logsigma parameters. For Normal2 and DoubleExp2, par.location and par.scale correspond
correspond to the location and scale parameters of the normal and double exponential
that are used to generates their squared forms.}
}
\description{
Random variates from several distributions are generated with user defined mean and variance.
}
\examples{
n = 5
dist = "Normal"
mu = 1
stdev = 0
par.location = 0
par.scale = 1
par.shape = 1
getDist(n=n, dist=dist, mu=mu, stdev=stdev, par.location = par.location, par.scale = par.scale, par.shape = par.shape)
}
