% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ObtainAverageRunLength.R
\name{getRL}
\alias{getRL}
\title{Run Length (RL)}
\usage{
getRL(replica = 1, n, m, theta = NULL, Ftheta = NULL, dist, mu,
  sigma, dist.par, chart, chart.par, calibrate = FALSE, arl0 = 370,
  alignment = "unadjusted", constant = NULL)
}
\arguments{
\item{replica}{is a numeric value. It is used for the parallel version of the function, it is set to 1 by default.}

\item{n}{is the subroup size}

\item{m}{is the reference sample size}

\item{theta}{is a constant.}

\item{Ftheta}{is a constant between (0,1).}

\item{dist}{is a vector of characters. Choose one type:
"Normal":     Normal distribution. (Default)
"DoubleExp":  Double exponential distribution (also known as Laplace distribution).
"DoubleExp2": Double exponential squared distribution from a DoubleExp(0,1).
"LogNormal":  Lognormal distribution.
"Gamma":      Gamma distribution.
"Weibull":    Weibull distribution}

\item{mu}{user defined expected value.}

\item{chart}{is the selected type of chart. Three options are available: Shewhart, CUSUM, EWMA}

\item{chart.par}{is a vector of one, or two elements depending
on the chart selected:
\describe{
  \item{For Shewhart scheme}{is c(k), where k comes from UCL = mu + k*sigma, LCL = mu - k*sigma.}
  \item{For CUSUM scheme}{is c(k, h, t) where k is the reference value and h is the control limit,
  and t is the type of the chart (1:positive, 2:negative, 3:two sides)}
  \item{For EWMA scheme}{is c(lambda, L), where lambda is the smoothing constant
  and L multiplies standard deviation to get the control limit}
}}

\item{calibrate}{is a boolean. If is TRUE the RL is limit to 10 times the target ARL}

\item{arl0}{is a numeric value. Is the expected value of the RL, by default is set to 370}

\item{alignment}{is the aligment of the data
\itemize{
  \item unadjusted:
  \item overallmean:
  \item overallmedian
  \item samplemean:
  \item samplemedian:
  \item referencemean:
  \item referencemedian:
  \item constantvalue:
}}

\item{constant}{is a numeric value}
}
\description{
Get the rung length
}
\examples{
n <- 5 # subgroup size
m <- 100 # reference-sample size
dist <- "Normal"
mu <- c(0, 0) # c(reference sample mean, monitoring sample mean)
sigma <- c(1, 1) # c(reference sample sd, monitoring sample sd)

#### Distribution parameters
dist.par <- c(0, 1, 1) # c(location, scale, shape)

#### Other Parameters
replicates <- 2
print.RL <- TRUE
calibrate <- FALSE
progress <- TRUE
arl0 <- 370

#### Control chart parameters
chart <- "Shewhart"
chart.par <- c(3)
shewhart <- getRL(1, n, m,
  theta = NULL, Ftheta = NULL, dist, mu, sigma, dist.par = dist.par,
  chart = chart, chart.par = chart.par, calibrate = calibrate, arl0 = arl0
)

chart <- "CUSUM"
chart.par <- c(0.25, 4.4181, 3)
cusum <- getRL(1, n, m,
  theta = NULL, Ftheta = NULL, dist, mu, sigma, dist.par = dist.par,
  chart = chart, chart.par = chart.par, calibrate = calibrate, arl0 = arl0
)

chart <- "EWMA"
chart.par <- c(0.2, 2.962)
shewhart <- getRL(1, n, m,
  theta = NULL, Ftheta = NULL, dist, mu, sigma, dist.par = dist.par,
  chart = chart, chart.par = chart.par, calibrate = calibrate, arl0 = arl0
)
}
