% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultivariateObtainAverageRunLength.R
\name{mGetRL}
\alias{mGetRL}
\title{Multivariate Run Length}
\usage{
mGetRL(replica = 1, n, m, nv, theta = NULL, Ftheta = NULL, dists, mu,
  sigma = NULL, dists.par = NULL, correlation = 0, chart = "T2",
  chart.par = c(10), alignment = "unadjusted", constant = NULL,
  absolute = FALSE, calibrate = FALSE, arl0 = 370)
}
\arguments{
\item{replica}{scalar. It is used for the parallel version of the function (\code{parallel=TRUE}). Default \code{1}.}

\item{n}{scalar. Subroup size}

\item{m}{scalar. Reference sample size}

\item{nv}{scalar. Number of variables to be generated.}

\item{theta}{vector. Value corresponig with the \code{Ftheta} quantile.}

\item{Ftheta}{vector. Quantile of the data distribution. The values that take are between (0,1).}

\item{dists}{vector of character string. Distribution of each variable. The length mus be the same as the number of variables. Select from:
\itemize{
  \item{"Normal": Normal distribution (default).}
  \item{"Gamma": Gamma distribution.}
}}

\item{mu}{scalar. Expected value of the desired distribution.}

\item{sigma}{scalar. Standard deviation of the desired distribution.}

\item{dists.par}{matrix  For each variable (column), specify
\itemize{
  \item{\code{par.location}: Location parameter of the desired distribution. Default 0.}
  \item{\code{par.scale}: Scale parameter of the desired distribution. Default 1.}
  \item{\code{par.shape}: Shape parameter of the desired distribution, Default 1.}
}
The number of columns must be the same as the number of variables.}

\item{correlation}{scalar. Corralation between variables.}

\item{chart}{character string. Selected type of chart. One option available: T2
\describe{
  \item{T2 scheme: }{is \code{c(k)}, where \code{k} comes from \eqn{UCL = mu + k\sigma, LCL = mu - k\sigma.}}
}}

\item{chart.par}{vector. Control limit and other parameters of the selected chart.}

\item{alignment}{character string. Aligment of the data \code{X} and \code{Y}. Select from
\itemize{
  \item "unadjusted": nothing is sustracte from \code{X} and \code{Y} (default).
  \item "overallmean": overall mean is sustracted from \code{X} and \code{Y}.
  \item "overallmedian": overall median is sustracted from \code{X} and \code{Y}.
  \item "samplemean": mean from corresponding group (\code{X} and \code{Y}) is sustracted from its corresponing vector.
  \item "samplemedian": median from corresponding group (\code{X} and \code{Y}) is sustracted from its corresponing vector.
  \item "referencemean": mean from \code{Y} is sustracted from \code{X} and \code{Y}.
  \item "referencemedian": median from \code{Y} is sustracted from \code{X} and \code{Y}.
  \item "constantvalue": a constant value is sustracted from \code{X} and \code{Y}.
}}

\item{constant}{scalar. Only used when the \code{alignment} is selected "constantvalue". Default \code{NULL}.}

\item{absolute}{logical. If \code{TRUE}, the absolute aligned values are obtained. (Default \code{FALSE})}

\item{calibrate}{logical. If \code{TRUE} the RL is limit to 10 times the target ARL.}

\item{arl0}{scalar. Expected value of the RL. Default \code{370}.}
}
\description{
Get the run length
}
\examples{
nv <- 2 #number of variables
inf <- 1000000 #infinite value to better approximation
alpha <- 0.005 #confindent interval
vec <- rchisq(inf, nv) #chi-sq random generator numbers according to the "infinite value"
h <- quantile(vec , 1-alpha) #control limit
mGetRL(n=5, m=10, nv=nv, mu=c(0,0), dists = c("Normal", "Normal"),
dists.par = matrix(c(0,1,1,0,1,1), ncol=2),
chart = "T2",chart.par=c(h), correlation=0)
}
