% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultivariateSequentialNormalScores.R
\name{MSNS}
\alias{MSNS}
\title{Multivariate Sequential Normal Scores}
\usage{
MSNS(X, X.id, Y = NULL, theta = NULL, Ftheta = NULL, scoring = "Z",
  alignment = "unadjusted", constant = NULL, absolute = FALSE,
  chart = "T2")
}
\arguments{
\item{X}{matrix or data.frame. New observations to obtain the N¡normal scores.}

\item{X.id}{vector. The id of each column (variable) of the matrix \code{X}.}

\item{Y}{matrix or data.frame. If \code{Y} is not defined (no previous observation available, \code{NULL}), NS is relative to \code{X}. Default \code{NULL}.}

\item{theta}{vector. Value corresponig with the \code{Ftheta} quantile.}

\item{Ftheta}{vector. Quantile of the data distribution. The values that take are between (0,1).}

\item{scoring}{character string. If "Z" (normal scores) (default). If "Z-SQ" (normal scores squared).}

\item{alignment}{character string. Aligment of the data \code{X} and \code{Y}. Select from
\itemize{
  \item "unadjusted": nothing is sustracte from \code{X} and \code{Y} (default).
  \item "overallmean": overall mean is sustracted from \code{X} and \code{Y}.
  \item "overallmedian": overall median is sustracted from \code{X} and \code{Y}.
  \item "samplemean": mean from corresponding group (\code{X} and \code{Y}) is sustracted from its corresponing vector.
  \item "samplemedian": median from corresponding group (\code{X} and \code{Y}) is sustracted from its corresponing vector.
  \item "referencemean": mean from \code{Y} is sustracted from \code{X} and \code{Y}.
  \item "referencemedian": median from \code{Y} is sustracted from \code{X} and \code{Y}.
  \item "constantvalue": a constant value is sustracted from \code{X} and \code{Y}.
}}

\item{constant}{scalar. Only used when the \code{alignment} is selected "constantvalue". Default \code{NULL}.}

\item{absolute}{logical. If \code{TRUE}, the absolute aligned values are obtained. (Default \code{FALSE})}

\item{chart}{character string.}
}
\description{
Transform a matrix \code{X} into SNS using initial observations \code{Y} if available
SNS follow the order of \code{X}.
}
\section{Comments}{

If ties, average ranks are used.
}

\examples{
X = cbind(example91$X1, example91$X2)
X.id = example91$X1.id
msns = MSNS(X, X.id)
}
\seealso{
\code{\link{MNS}} for multivariate normal scores
}
