% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultivariateSequentialNormalScores.R
\name{plot.msns}
\alias{plot.msns}
\title{Plot Multivariate Sequential Normal Scores}
\usage{
\method{plot}{msns}(object, ...)
}
\arguments{
\item{object}{msns class.}

\item{...}{same parameters for \code{plot} function.}
}
\description{
plot the Multivariate Sequential Normal Scores by using only \code{plot}
}
