% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ObtainAverageRunLength.R
\name{getRL}
\alias{getRL}
\title{Run Length}
\usage{
getRL(replica = 1, n, m, theta = NULL, Ftheta = NULL, dist, mu,
  sigma, dist.par = c(0, 1, 1), chart, chart.par, calibrate = FALSE,
  arl0 = 370, alignment = "unadjusted", constant = NULL,
  absolute = FALSE)
}
\arguments{
\item{replica}{scalar. It is used for the parallel version of the function (\code{parallel=TRUE}). Default \code{1}.}

\item{n}{scalar. Subroup size}

\item{m}{scalar. Reference sample size}

\item{theta}{scalar. Value corresponig with the \code{Ftheta} quantile.}

\item{Ftheta}{scalar. Quantile of the data distribution. The values that take are between (0,1).}

\item{dist}{character string. Select from:
\itemize{
  \item{"Normal": Normal distribution (default).}
  \item{"DoubleExp": Double exponential distribution (also known as Laplace distribution).}
  \item{"DoubleExp2": Double exponential squared distribution from a \code{DoubleExp(0,1)}.}
  \item{"LogNormal": Lognormal distribution.}
  \item{"Gamma": Gamma distribution.}
  \item{"Weibull": Weibull distribution.}
}}

\item{mu}{scalar. Expected value of the desired distribution.}

\item{sigma}{scalar. Standard deviation of the desired distribution.}

\item{dist.par}{vector. Distribution parameters. \code{c(par.location, par.scale, par.shape)}. Default \code{c(0,1,1)}.}

\item{chart}{character string. Selected type of chart. Three options are available: Shewhart, CUSUM, EWMA}

\item{chart.par}{vector. The size depends on the selected chart:
\describe{
  \item{Shewhart scheme: }{is \code{c(k)}, where \code{k} comes from \eqn{UCL = mu + k\sigma, LCL = mu - k\sigma.}}
  \item{CUSUM scheme: }{is \code{c(k, h, t)} where \code{k} is the reference value and \code{h} is the control limit,
  and \code{t} is the type of the chart (1:positive, 2:negative, 3:two sides)}
  \item{EWMA scheme: }{is \code{c(lambda, L)}, where \code{lambda} is the smoothing constant
  and \code{L} multiplies standard deviation to get the control limit}
}}

\item{calibrate}{logical. If \code{TRUE} the RL is limit to 10 times the target ARL.}

\item{arl0}{scalar. Expected value of the RL. Default \code{370}.}

\item{alignment}{character string. Aligment of the data \code{X} and \code{Y}. Select from
\itemize{
  \item "unadjusted": nothing is sustracte from \code{X} and \code{Y} (default).
  \item "overallmean": overall mean is sustracted from \code{X} and \code{Y}.
  \item "overallmedian": overall median is sustracted from \code{X} and \code{Y}.
  \item "samplemean": mean from corresponding group (\code{X} and \code{Y}) is sustracted from its corresponing vector.
  \item "samplemedian": median from corresponding group (\code{X} and \code{Y}) is sustracted from its corresponing vector.
  \item "referencemean": mean from \code{Y} is sustracted from \code{X} and \code{Y}.
  \item "referencemedian": median from \code{Y} is sustracted from \code{X} and \code{Y}.
  \item "constantvalue": a constant value is sustracted from \code{X} and \code{Y}.
}}

\item{constant}{scalar. Only used when the \code{alignment} is selected "constantvalue". Default \code{NULL}.}

\item{absolute}{logical. If \code{TRUE}, the absolute aligned values are obtained. (Default \code{FALSE})}
}
\description{
Get the run length
}
\examples{
n <- 5 # subgroup size
m <- 100 # reference-sample size
dist <- "Normal"
mu <- c(0, 0) # c(reference sample mean, monitoring sample mean)
sigma <- c(1, 1) # c(reference sample sd, monitoring sample sd)

#### Distribution parameters
dist.par <- c(0, 1, 1) # c(location, scale, shape)

#### Other Parameters
replicates <- 2
print.RL <- TRUE
calibrate <- FALSE
progress <- TRUE
arl0 <- 370

#### Control chart parameters
chart <- "Shewhart"
chart.par <- c(3)
shewhart <- getRL(1, n, m,
  theta = NULL, Ftheta = NULL, dist, mu, sigma, dist.par = dist.par,
  chart = chart, chart.par = chart.par, calibrate = calibrate, arl0 = arl0
)

chart <- "CUSUM"
chart.par <- c(0.25, 4.4181, 3)
cusum <- getRL(1, n, m,
  theta = NULL, Ftheta = NULL, dist, mu, sigma, dist.par = dist.par,
  chart = chart, chart.par = chart.par, calibrate = calibrate, arl0 = arl0
)

chart <- "EWMA"
chart.par <- c(0.2, 2.962)
shewhart <- getRL(1, n, m,
  theta = NULL, Ftheta = NULL, dist, mu, sigma, dist.par = dist.par,
  chart = chart, chart.par = chart.par, calibrate = calibrate, arl0 = arl0
)
}
