% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NS.R
\name{NS}
\alias{NS}
\title{Normal scores X from Y}
\usage{
NS(X, Y = NULL, theta = NULL, Ftheta = NULL)
}
\arguments{
\item{X}{is a numerical vector.}

\item{Y}{is a numerical vector. If Y is not defined, by default Y = NULL.}

\item{theta}{is a constant.}

\item{Ftheta}{is a constant between (0,1).}
}
\value{
Z (Normal Scores)
}
\description{
Get conditional and unconditional normal score (NS) of observations (X)
relative to previous observations (Y).
}
\details{
If Y = NULL (no previous observation available), NS is relative to X.
GENERAL COMMENTS
If ties, average ranks are used.
If Y = NULL, normal scores are set relative to X.
UNCONDITIONAL COMMENTS
Instead of Van Der Waerden Normal Scores where p = r/(n+1), p = (r-0.5)/n,
where r stands for rank and p for the input evaluated in the
inverse of a Standard Normal Distribution.
}
\examples{
Y = c(10,20,30,40,50,60,70,80,90,100)
X = c(30, 35, 45)
theta = 40
Ftheta = 0.5
#EXAMPLE CONDITIONAL
NS(X = X, Y = Y, theta = theta, Ftheta = Ftheta)
#[1] -0.52440051 -0.38532047  0.08964235
#EXAMPLE UNCONDITIONAL
theta = NULL
Ftheta = NULL
NS(X = X, Y = Y, theta = theta, Ftheta = Ftheta)
#[1] -0.6045853 -0.4727891 -0.2298841
}
