% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SequentialNormalScores.R
\name{plot.sns}
\alias{plot.sns}
\title{Plot Sequential Normal Scores}
\usage{
\method{plot}{sns}(object, ...)
}
\arguments{
\item{object}{sns class.}

\item{...}{same parameters for \code{plot} function.}
}
\description{
plot the Sequential Normal Scores by using only \code{plot}
}
