% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNS.R
\name{SNS}
\alias{SNS}
\title{Sequential normal scores}
\usage{
SNS(X, X.id, Y = NULL, theta = NULL, Ftheta = NULL)
}
\arguments{
\item{X}{is a numerical vector.}

\item{X.id}{is the id of the vector X}

\item{Y}{is a numerical vector. Y = NULL if undefined.}

\item{theta}{is a constant}

\item{Ftheta}{is a constant between (0,1)}
}
\description{
Transform a vector X into SNS using initial observations Y if available
SNS follow the order of X.
}
\details{
If ties, average ranks are used.
If Y = NULL, normal scores are set relative to X.
}
\examples{
#EXAMPLE CONDITIONAL WITH REFERENCE SAMPLE
Y = c(10,20,30,40,50,60,70,80,90,100)
X = c(30, 35, 45)
theta = 40
Ftheta = 0.5
sample.id = c("a", "b", "c")
SNS(X=X, X.id=sample.id, Y = Y, theta = theta, Ftheta = Ftheta)
#[1] -0.52440051 -0.31863936  0.08964235

#EXAMPLE CONDITIONAL WITH REFERENCE SAMPLE
Y = c(10,20,30,40,50,60,70,80,90,100)
X = c(30, 35, 45)
theta = 40
Ftheta = 0.5
sample.id = c("a", "b", "c")
SNS(X=X, X.id=sample.id, Y = Y, theta = theta, Ftheta = Ftheta)
#[1] -0.52440051 -0.31863936  0.08964235

#EXAMPLE UNCONDITIONAL WITH REFERENCE SAMPLE
Y = c(10,20,30,40,50,60,70,80,90,100)
X = c(30, 35, 45)
theta = NULL
Ftheta = NULL
sample.id = c("a", "b", "c")
SNS(X=X, X.id=sample.id, Y = Y, theta = theta, Ftheta = Ftheta)
#[1] -0.6045853 -0.3186394  0.0000000

#EXAMPLE CONDITIONAL WITHOUT REFERENCE SAMPLE
Y = NULL#c(10,20,30,40,50,60,70,80,90,100)
X = c(30, 35, 45)
theta = 40
Ftheta = 0.5
sample.id = c("a", "b", "c")
SNS(X=X, X.id=sample.id, Y = Y, theta = theta, Ftheta = Ftheta)
#[1] -0.6744898 -0.3186394  0.6744898

#EXAMPLE UNCONDITIONAL WITHOUT REFERENCE SAMPLE
Y = NULL
X = c(30, 35, 45)
theta = NULL
Ftheta = NULL
sample.id = c("a", "b", "c")
SNS(X=X, X.id=sample.id, Y = Y, theta = theta, Ftheta = Ftheta)
#[1] 0.0000000 0.6744898 0.9674216
}
